refactor add remove links and group by to a separate file
