+++ packages/nocodb/src/db/BaseModelSqlv2.ts
@@ -32,6 +32,7 @@ import {
 } from 'nocodb-sdk';
 import { v4 as uuidv4 } from 'uuid';
 import Validator from 'validator';
+import { addOrRemoveLinks } from './BaseModelSqlv2/add-remove-links';
 import { NestedLinkPreparator } from './BaseModelSqlv2/nested-link-preparator';
 import { relationDataFetcher } from './BaseModelSqlv2/relation-data-fetcher';
 import type { Knex } from 'knex';
@@ -70,9 +71,9 @@ import type {
 import type LookupColumn from '~/models/LookupColumn';
 import type { ResolverObj } from '~/utils';
 import applyAggregation from '~/db/aggregation';
+import { groupBy as baseModelGroupBy } from '~/db/BaseModelSqlv2/group-by';
 import conditionV2 from '~/db/conditionV2';
 import formulaQueryBuilderv2 from '~/db/formulav2/formulaQueryBuilderv2';
-import generateLookupSelectQuery from '~/db/generateLookupSelectQuery';
 import genRollupSelectv2 from '~/db/genRollupSelectv2';
 import { RelationManager } from '~/db/relation-manager';
 import sortV2 from '~/db/sortV2';
@@ -82,15 +83,12 @@ import {
   _wherePk,
   applyPaginate,
   checkColumnRequired,
-  extractIds,
   extractSortsObject,
   formatDataForAudit,
   getAs,
   getColumnName,
   getCompositePkValue,
   getListArgs,
-  getOppositeRelationType,
-  getRelatedLinksColumn,
   haveFormulaColumn,
   isDataAuditEnabled as isDataAuditEnabledFn,
   isPrimitiveType,
@@ -122,7 +120,6 @@ import {
   extractColsMetaForAudit,
   extractExcludedColumnNames,
   generateAuditV1Payload,
-  getAliasGenerator,
   nocoExecute,
   populateUpdatePayloadDiff,
   remapWithAlias,
@@ -151,7 +148,10 @@ const MAX_RECURSION_DEPTH = 2;
  */
 class BaseModelSqlv2 implements IBaseModelSqlV2 {
   protected _dbDriver: XKnex;
-  protected viewId: string;
+  protected _viewId: string;
+  public get viewId() {
+    return this._viewId;
+  }
   protected _proto: any;
   protected _columns = {};
   protected source: Source;
@@ -178,7 +178,7 @@ class BaseModelSqlv2 implements IBaseModelSqlV2 {
   }) {
     this._dbDriver = dbDriver;
     this.model = model;
-    this.viewId = viewId;
+    this._viewId = viewId;
     this.context = context;
     this.schema = schema;
     autoBind(this);
@@ -767,299 +767,11 @@ class BaseModelSqlv2 implements IBaseModelSqlV2 {
     }[],
     _view: View,
   ) {
-    try {
-      const columns = await this.model.getColumns(this.context);
-      const aliasColObjMap = await this.model.getAliasColObjMap(
-        this.context,
-        columns,
-      );
-      const selectors = [] as Array<Knex.Raw>;
-
-      const viewFilterList = await Filter.rootFilterList(this.context, {
-        viewId: this.viewId,
-      });
-
-      if (!bulkFilterList?.length) {
-        return NcError.badRequest('bulkFilterList is required');
-      }
-
-      for (const f of bulkFilterList) {
-        const { where, ...rest } = this._getListArgs(f);
-        const groupBySelectors = [];
-        const groupByColumns: Record<string, Column> = {};
-
-        const getAlias = getAliasGenerator('__nc_gb');
-        const { filters: groupFilter } = extractFilterFromXwhere(
-          this.context,
-          f.where,
-          aliasColObjMap,
-        );
-
-        const tQb = this.dbDriver(this.tnPath);
-        const colSelectors = [];
-
-        await Promise.all(
-          rest.column_name.split(',').map(async (col) => {
-            let column = columns.find(
-              (c) => c.column_name === col || c.title === col,
-            );
-
-            // if qrCode or Barcode replace it with value column nd keep the alias
-            if ([UITypes.QrCode, UITypes.Barcode].includes(column.uidt)) {
-              column = new Column({
-                ...(await column
-                  .getColOptions<BarcodeColumn | QrCodeColumn>(this.context)
-                  .then((col) => col.getValueColumn(this.context))),
-                asId: column.id,
-              });
-            }
-
-            groupByColumns[getAs(column)] = column;
-
-            switch (column.uidt) {
-              case UITypes.Attachment:
-                NcError.badRequest(
-                  'Group by using attachment column is not supported',
-                );
-                break;
-              case UITypes.Button: {
-                NcError.badRequest(
-                  'Group by using Button column is not supported',
-                );
-                break;
-              }
-              case UITypes.Links:
-              case UITypes.Rollup:
-                colSelectors.push(
-                  (
-                    await genRollupSelectv2({
-                      baseModelSqlv2: this,
-                      knex: this.dbDriver,
-                      columnOptions: (await column.getColOptions(
-                        this.context,
-                      )) as RollupColumn,
-                    })
-                  ).builder.as(getAs(column)),
-                );
-                groupBySelectors.push(getAs(column));
-                break;
-              case UITypes.Formula: {
-                let selectQb;
-                try {
-                  const _selectQb = await this.getSelectQueryBuilderForFormula(
-                    column,
-                  );
-                  selectQb = this.dbDriver.raw(`?? as ??`, [
-                    _selectQb.builder,
-                    getAs(column),
-                  ]);
-                } catch (e) {
-                  console.log(e);
-                  selectQb = this.dbDriver.raw(`'ERR' as ??`, [getAs(column)]);
-                }
-                colSelectors.push(selectQb);
-                groupBySelectors.push(getAs(column));
-                break;
-              }
-
-              case UITypes.Lookup:
-              case UITypes.LinkToAnotherRecord: {
-                const _selectQb = await generateLookupSelectQuery({
-                  baseModelSqlv2: this,
-                  column,
-                  alias: null,
-                  model: this.model,
-                  getAlias,
-                });
-                const selectQb = this.dbDriver.raw(`?? as ??`, [
-                  this.dbDriver.raw(_selectQb.builder).wrap('(', ')'),
-                  getAs(column),
-                ]);
-                colSelectors.push(selectQb);
-                groupBySelectors.push(getAs(column));
-                break;
-              }
-              case UITypes.DateTime:
-              case UITypes.CreatedTime:
-              case UITypes.LastModifiedTime:
-                {
-                  const columnName = await getColumnName(
-                    this.context,
-                    column,
-                    columns,
-                  );
-                  // ignore seconds part in datetime and group
-                  if (this.dbDriver.clientType() === 'pg') {
-                    colSelectors.push(
-                      this.dbDriver.raw(
-                        "date_trunc('minute', ??) + interval '0 seconds' as ??",
-                        [columnName, getAs(column)],
-                      ),
-                    );
-                  } else if (
-                    this.dbDriver.clientType() === 'mysql' ||
-                    this.dbDriver.clientType() === 'mysql2'
-                  ) {
-                    colSelectors.push(
-                      // this.dbDriver.raw('??::date as ??', [columnName, getAs(column)]),
-                      this.dbDriver.raw(
-                        "DATE_SUB(CONVERT_TZ(??, @@GLOBAL.time_zone, '+00:00'), INTERVAL SECOND(??) SECOND) as ??",
-                        [columnName, columnName, getAs(column)],
-                      ),
-                    );
-                  } else if (this.dbDriver.clientType() === 'sqlite3') {
-                    colSelectors.push(
-                      this.dbDriver.raw(
-                        `strftime ('%Y-%m-%d %H:%M:00',:column:) ||
-  (
-   CASE WHEN substr(:column:, 20, 1) = '+' THEN
-    printf ('+%s:',
-     substr(:column:, 21, 2)) || printf ('%s',
-     substr(:column:, 24, 2))
-   WHEN substr(:column:, 20, 1) = '-' THEN
-    printf ('-%s:',
-     substr(:column:, 21, 2)) || printf ('%s',
-     substr(:column:, 24, 2))
-   ELSE
-    '+00:00'
-   END) AS :id:`,
-                        {
-                          column: columnName,
-                          id: getAs(column),
-                        },
-                      ),
-                    );
-                  } else {
-                    colSelectors.push(
-                      this.dbDriver.raw('DATE(??) as ??', [
-                        columnName,
-                        getAs(column),
-                      ]),
-                    );
-                  }
-                  groupBySelectors.push(getAs(column));
-                }
-                break;
-              default: {
-                const columnName = await getColumnName(
-                  this.context,
-                  column,
-                  columns,
-                );
-                colSelectors.push(
-                  this.dbDriver.raw('?? as ??', [columnName, getAs(column)]),
-                );
-                groupBySelectors.push(getAs(column));
-                break;
-              }
-            }
-          }),
-        );
-
-        // get aggregated count of each group
-        tQb.count(`${this.model.primaryKey?.column_name || '*'} as count`);
-        tQb.select(...colSelectors);
-
-        if (+rest?.shuffle) {
-          await this.shuffle({ qb: tQb });
-        }
-
-        await conditionV2(
-          this,
-          [
-            ...(this.viewId
-              ? [
-                  new Filter({
-                    children: viewFilterList || [],
-                    is_group: true,
-                  }),
-                ]
-              : []),
-            new Filter({
-              children: rest.filterArr || [],
-              is_group: true,
-              logical_op: 'and',
-            }),
-            new Filter({
-              children: extractFilterFromXwhere(
-                this.context,
-                where,
-                aliasColObjMap,
-              ).filters,
-              is_group: true,
-              logical_op: 'and',
-            }),
-            new Filter({
-              children: groupFilter,
-              is_group: true,
-              logical_op: 'and',
-            }),
-            new Filter({
-              children: args.filterArr || [],
-              is_group: true,
-              logical_op: 'and',
-            }),
-          ],
-          tQb,
-        );
-
-        tQb.groupBy(...groupBySelectors);
-
-        const count = this.dbDriver
-          .count('*', { as: 'count' })
-          .from(tQb.as('groupby'));
-
-        let subQuery;
-        switch (this.dbDriver.client.config.client) {
-          case 'pg':
-            subQuery = this.dbDriver
-              .select(
-                this.dbDriver.raw(`json_build_object('count', "count") as ??`, [
-                  getAlias(),
-                ]),
-              )
-              .from(count.as(getAlias()));
-            selectors.push(
-              this.dbDriver.raw(`(??) as ??`, [subQuery, `${f.alias}`]),
-            );
-            break;
-          case 'mysql2':
-            subQuery = this.dbDriver
-              .select(this.dbDriver.raw(`JSON_OBJECT('count', \`count\`)`))
-              .from(count.as(getAlias()));
-            selectors.push(
-              this.dbDriver.raw(`(??) as ??`, [subQuery, `${f.alias}`]),
-            );
-            break;
-          case 'sqlite3':
-            subQuery = this.dbDriver
-              .select(
-                this.dbDriver.raw(`json_object('count', "count") as ??`, [
-                  f.alias,
-                ]),
-              )
-              .from(count.as(getAlias()));
-            selectors.push(
-              this.dbDriver.raw(`(??) as ??`, [subQuery, `${f.alias}`]),
-            );
-            break;
-          default:
-            NcError.notImplemented(
-              'This database does not support bulk groupBy count',
-            );
-        }
-      }
-
-      const qb = this.dbDriver(this.tnPath);
-      qb.select(...selectors).limit(1);
-
-      return await this.execAndParse(qb, null, {
-        raw: true,
-        first: true,
-      });
-    } catch (e) {
-      console.log(e);
-    }
+    return await baseModelGroupBy(this, logger).bulkCount(
+      args,
+      bulkFilterList,
+      _view,
+    );
   }
 
   async bulkGroupBy(
@@ -1078,383 +790,11 @@ class BaseModelSqlv2 implements IBaseModelSqlV2 {
     }[],
     _view: View,
   ) {
-    const columns = await this.model.getColumns(this.context);
-    const aliasColObjMap = await this.model.getAliasColObjMap(
-      this.context,
-      columns,
+    return await baseModelGroupBy(this, logger).bulkList(
+      args,
+      bulkFilterList,
+      _view,
     );
-    const selectors = [] as Array<Knex.Raw>;
-
-    const viewFilterList = await Filter.rootFilterList(this.context, {
-      viewId: this.viewId,
-    });
-
-    try {
-      if (!bulkFilterList?.length) {
-        return NcError.badRequest('bulkFilterList is required');
-      }
-
-      for (const f of bulkFilterList) {
-        const { where, ...rest } = this._getListArgs(f);
-        const groupBySelectors = [];
-        const groupByColumns: Record<string, Column> = {};
-
-        const getAlias = getAliasGenerator('__nc_gb');
-        const { filters: groupFilter } = extractFilterFromXwhere(
-          this.context,
-          f?.where,
-          aliasColObjMap,
-        );
-        let groupSort = extractSortsObject(rest?.sort, aliasColObjMap);
-
-        const tQb = this.dbDriver(this.tnPath);
-        const colSelectors = [];
-        const colIds = rest.column_name
-          .split(',')
-          .map((col) => {
-            const column = columns.find(
-              (c) => c.column_name === col || c.title === col,
-            );
-            if (!column) {
-              throw NcError.fieldNotFound(col);
-            }
-            return column?.id;
-          })
-          .join('_');
-
-        await Promise.all(
-          rest.column_name.split(',').map(async (col) => {
-            let column = columns.find(
-              (c) => c.column_name === col || c.title === col,
-            );
-
-            // if qrCode or Barcode replace it with value column nd keep the alias
-            if ([UITypes.QrCode, UITypes.Barcode].includes(column.uidt)) {
-              column = new Column({
-                ...(await column
-                  .getColOptions<BarcodeColumn | QrCodeColumn>(this.context)
-                  .then((col) => col.getValueColumn(this.context))),
-                asId: column.id,
-              });
-            }
-
-            groupByColumns[getAs(column)] = column;
-
-            switch (column.uidt) {
-              case UITypes.Attachment:
-                NcError.badRequest(
-                  'Group by using attachment column is not supported',
-                );
-                break;
-              case UITypes.Button: {
-                NcError.badRequest(
-                  'Group by using Button column is not supported',
-                );
-                break;
-              }
-              case UITypes.Links:
-              case UITypes.Rollup:
-                colSelectors.push(
-                  (
-                    await genRollupSelectv2({
-                      baseModelSqlv2: this,
-                      knex: this.dbDriver,
-                      columnOptions: (await column.getColOptions(
-                        this.context,
-                      )) as RollupColumn,
-                    })
-                  ).builder.as(getAs(column)),
-                );
-                groupBySelectors.push(getAs(column));
-                break;
-              case UITypes.Formula: {
-                let selectQb;
-                try {
-                  const _selectQb = await this.getSelectQueryBuilderForFormula(
-                    column,
-                  );
-                  selectQb = this.dbDriver.raw(`?? as ??`, [
-                    _selectQb.builder,
-                    getAs(column),
-                  ]);
-                } catch (e) {
-                  console.log(e);
-                  selectQb = this.dbDriver.raw(`'ERR' as ??`, [getAs(column)]);
-                }
-                colSelectors.push(selectQb);
-                groupBySelectors.push(getAs(column));
-                break;
-              }
-
-              case UITypes.Lookup:
-              case UITypes.LinkToAnotherRecord: {
-                const _selectQb = await generateLookupSelectQuery({
-                  baseModelSqlv2: this,
-                  column,
-                  alias: null,
-                  model: this.model,
-                  getAlias,
-                });
-                const selectQb = this.dbDriver.raw(`?? as ??`, [
-                  this.dbDriver.raw(_selectQb.builder).wrap('(', ')'),
-                  getAs(column),
-                ]);
-                colSelectors.push(selectQb);
-                groupBySelectors.push(getAs(column));
-                break;
-              }
-              case UITypes.DateTime:
-              case UITypes.CreatedTime:
-              case UITypes.LastModifiedTime:
-                {
-                  const columnName = await getColumnName(
-                    this.context,
-                    column,
-                    columns,
-                  );
-                  // ignore seconds part in datetime and group
-                  if (this.dbDriver.clientType() === 'pg') {
-                    colSelectors.push(
-                      this.dbDriver.raw(
-                        "date_trunc('minute', ??) + interval '0 seconds' as ??",
-                        [columnName, getAs(column)],
-                      ),
-                    );
-                  } else if (
-                    this.dbDriver.clientType() === 'mysql' ||
-                    this.dbDriver.clientType() === 'mysql2'
-                  ) {
-                    colSelectors.push(
-                      // this.dbDriver.raw('??::date as ??', [columnName, getAs(column)]),
-                      this.dbDriver.raw(
-                        "DATE_SUB(CONVERT_TZ(??, @@GLOBAL.time_zone, '+00:00'), INTERVAL SECOND(??) SECOND) as ??",
-                        [columnName, columnName, getAs(column)],
-                      ),
-                    );
-                  } else if (this.dbDriver.clientType() === 'sqlite3') {
-                    colSelectors.push(
-                      this.dbDriver.raw(
-                        `strftime ('%Y-%m-%d %H:%M:00',:column:) ||
-  (
-   CASE WHEN substr(:column:, 20, 1) = '+' THEN
-    printf ('+%s:',
-     substr(:column:, 21, 2)) || printf ('%s',
-     substr(:column:, 24, 2))
-   WHEN substr(:column:, 20, 1) = '-' THEN
-    printf ('-%s:',
-     substr(:column:, 21, 2)) || printf ('%s',
-     substr(:column:, 24, 2))
-   ELSE
-    '+00:00'
-   END) AS :id:`,
-                        {
-                          column: columnName,
-                          id: getAs(column),
-                        },
-                      ),
-                    );
-                  } else {
-                    colSelectors.push(
-                      this.dbDriver.raw('DATE(??) as ??', [
-                        columnName,
-                        getAs(column),
-                      ]),
-                    );
-                  }
-                  groupBySelectors.push(getAs(column));
-                }
-                break;
-              default: {
-                const columnName = await getColumnName(
-                  this.context,
-                  column,
-                  columns,
-                );
-                colSelectors.push(
-                  this.dbDriver.raw('?? as ??', [columnName, getAs(column)]),
-                );
-                groupBySelectors.push(getAs(column));
-                break;
-              }
-            }
-          }),
-        );
-
-        // get aggregated count of each group
-        tQb.count(`${this.model.primaryKey?.column_name || '*'} as count`);
-        tQb.select(...colSelectors);
-
-        if (+rest?.shuffle) {
-          await this.shuffle({ qb: tQb });
-        }
-
-        await conditionV2(
-          this,
-          [
-            ...(this.viewId
-              ? [
-                  new Filter({
-                    children: viewFilterList || [],
-                    is_group: true,
-                  }),
-                ]
-              : []),
-            new Filter({
-              children: rest.filterArr || [],
-              is_group: true,
-              logical_op: 'and',
-            }),
-            new Filter({
-              children: extractFilterFromXwhere(
-                this.context,
-                where,
-                aliasColObjMap,
-              ).filters,
-              is_group: true,
-              logical_op: 'and',
-            }),
-            new Filter({
-              children: groupFilter,
-              is_group: true,
-              logical_op: 'and',
-            }),
-            new Filter({
-              children: args.filterArr || [],
-              is_group: true,
-              logical_op: 'and',
-            }),
-          ],
-          tQb,
-        );
-
-        if (!groupSort) {
-          if (rest.sortArr?.length) {
-            groupSort = rest.sortArr;
-          } else if (this.viewId) {
-            groupSort = await Sort.list(this.context, { viewId: this.viewId });
-          }
-        }
-
-        for (const sort of groupSort || []) {
-          if (!groupByColumns[sort.fk_column_id]) {
-            continue;
-          }
-
-          const column = groupByColumns[sort.fk_column_id];
-
-          if (
-            [UITypes.User, UITypes.CreatedBy, UITypes.LastModifiedBy].includes(
-              column.uidt as UITypes,
-            )
-          ) {
-            const columnName = await getColumnName(
-              this.context,
-              column,
-              columns,
-            );
-            const baseUsers = await BaseUser.getUsersList(this.context, {
-              base_id: column.base_id,
-            });
-
-            // create nested replace statement for each user
-            const finalStatement = baseUsers.reduce((acc, user) => {
-              const qb = this.dbDriver.raw(`REPLACE(${acc}, ?, ?)`, [
-                user.id,
-                user.display_name || user.email,
-              ]);
-              return qb.toQuery();
-            }, this.dbDriver.raw(`??`, [columnName]).toQuery());
-
-            if (!['asc', 'desc'].includes(sort.direction)) {
-              tQb.orderBy(
-                'count',
-                sort.direction === 'count-desc' ? 'desc' : 'asc',
-                sort.direction === 'count-desc' ? 'LAST' : 'FIRST',
-              );
-            } else {
-              tQb.orderBy(
-                sanitize(this.dbDriver.raw(finalStatement)),
-                sort.direction,
-                sort.direction === 'desc' ? 'LAST' : 'FIRST',
-              );
-            }
-          } else {
-            if (!['asc', 'desc'].includes(sort.direction)) {
-              tQb.orderBy(
-                'count',
-                sort.direction === 'count-desc' ? 'desc' : 'asc',
-                sort.direction === 'count-desc' ? 'LAST' : 'FIRST',
-              );
-            } else {
-              tQb.orderBy(
-                getAs(column),
-                sort.direction,
-                sort.direction === 'desc' ? 'LAST' : 'FIRST',
-              );
-            }
-          }
-          tQb.groupBy(...groupBySelectors);
-          applyPaginate(tQb, rest);
-        }
-
-        let subQuery;
-        switch (this.dbDriver.client.config.client) {
-          case 'pg':
-            subQuery = this.dbDriver
-              .select(
-                this.dbDriver.raw(
-                  `json_agg(json_build_object('count', "count", '${rest.column_name}', "${colIds}")) as ??`,
-                  [getAlias()],
-                ),
-              )
-              .from(tQb.as(getAlias()));
-            selectors.push(
-              this.dbDriver.raw(`(??) as ??`, [subQuery, `${f.alias}`]),
-            );
-            break;
-          case 'mysql2':
-            subQuery = this.dbDriver
-              .select(
-                this.dbDriver.raw(
-                  `JSON_ARRAYAGG(JSON_OBJECT('count', \`count\`, '${rest.column_name}', \`${colIds}\`))`,
-                ),
-              )
-              .from(this.dbDriver.raw(`(??) as ??`, [tQb, getAlias()]));
-            selectors.push(
-              this.dbDriver.raw(`(??) as ??`, [subQuery, f.alias]),
-            );
-            break;
-          case 'sqlite3':
-            subQuery = this.dbDriver
-              .select(
-                this.dbDriver.raw(
-                  `json_group_array(json_object('count', "count", '${rest.column_name}', "${colIds}")) as ??`,
-                  [f.alias],
-                ),
-              )
-              .from(tQb.as(getAlias()));
-            selectors.push(
-              this.dbDriver.raw(`(??) as ??`, [subQuery, f.alias]),
-            );
-            break;
-          default:
-            NcError.notImplemented(
-              'This database does not support bulk groupBy',
-            );
-        }
-      }
-
-      const qb = this.dbDriver(this.tnPath);
-      qb.select(...selectors).limit(1);
-
-      return await this.execAndParse(qb, null, {
-        raw: true,
-        first: true,
-      });
-    } catch (err) {
-      logger.log(err);
-      return [];
-    }
   }
 
   async bulkAggregate(
@@ -1784,310 +1124,7 @@ class BaseModelSqlv2 implements IBaseModelSqlV2 {
     filterArr?: Filter[];
     sortArr?: Sort[];
   }) {
-    const { where, ...rest } = this._getListArgs(args as any);
-
-    args.column_name = args.column_name || '';
-
-    const columns = await this.model.getColumns(this.context);
-    const groupByColumns: Record<string, Column> = {};
-
-    const selectors = [];
-    const groupBySelectors = [];
-    const getAlias = getAliasGenerator('__nc_gb');
-
-    await Promise.all(
-      args.column_name.split(',').map(async (col) => {
-        let column = columns.find(
-          (c) => c.column_name === col || c.title === col,
-        );
-        if (!column) {
-          throw NcError.fieldNotFound(col);
-        }
-
-        // if qrCode or Barcode replace it with value column nd keep the alias
-        if ([UITypes.QrCode, UITypes.Barcode].includes(column.uidt))
-          column = new Column({
-            ...(await column
-              .getColOptions<BarcodeColumn | QrCodeColumn>(this.context)
-              .then((col) => col.getValueColumn(this.context))),
-            asId: column.id,
-          });
-
-        groupByColumns[getAs(column)] = column;
-
-        switch (column.uidt) {
-          case UITypes.Attachment:
-            NcError.badRequest(
-              'Group by using attachment column is not supported',
-            );
-            break;
-          case UITypes.Button:
-            {
-              NcError.badRequest(
-                'Group by using Button column is not supported',
-              );
-            }
-            break;
-          case UITypes.Links:
-          case UITypes.Rollup:
-            selectors.push(
-              (
-                await genRollupSelectv2({
-                  baseModelSqlv2: this,
-                  knex: this.dbDriver,
-                  columnOptions: (await column.getColOptions(
-                    this.context,
-                  )) as RollupColumn,
-                })
-              ).builder.as(getAs(column)),
-            );
-            groupBySelectors.push(getAs(column));
-            break;
-          case UITypes.Formula:
-            {
-              let selectQb;
-              try {
-                const _selectQb = await this.getSelectQueryBuilderForFormula(
-                  column,
-                );
-
-                selectQb = this.dbDriver.raw(`?? as ??`, [
-                  _selectQb.builder,
-                  getAs(column),
-                ]);
-              } catch (e) {
-                logger.log(e);
-                // return dummy select
-                selectQb = this.dbDriver.raw(`'ERR' as ??`, [getAs(column)]);
-              }
-
-              selectors.push(selectQb);
-              groupBySelectors.push(getAs(column));
-            }
-            break;
-          case UITypes.Lookup:
-          case UITypes.LinkToAnotherRecord:
-            {
-              const _selectQb = await generateLookupSelectQuery({
-                baseModelSqlv2: this,
-                column,
-                alias: null,
-                model: this.model,
-                getAlias,
-              });
-
-              const selectQb = this.dbDriver.raw(`?? as ??`, [
-                this.dbDriver.raw(_selectQb.builder).wrap('(', ')'),
-                getAs(column),
-              ]);
-
-              selectors.push(selectQb);
-              groupBySelectors.push(getAs(column));
-            }
-            break;
-          case UITypes.CreatedTime:
-          case UITypes.LastModifiedTime:
-          case UITypes.DateTime:
-            {
-              const columnName = await getColumnName(
-                this.context,
-                column,
-                columns,
-              );
-              // ignore seconds part in datetime and group
-              if (this.dbDriver.clientType() === 'pg') {
-                selectors.push(
-                  this.dbDriver.raw(
-                    "date_trunc('minute', ??) + interval '0 seconds' as ??",
-                    [columnName, getAs(column)],
-                  ),
-                );
-              } else if (
-                this.dbDriver.clientType() === 'mysql' ||
-                this.dbDriver.clientType() === 'mysql2'
-              ) {
-                selectors.push(
-                  // this.dbDriver.raw('??::date as ??', [columnName, getAs(column)]),
-                  this.dbDriver.raw(
-                    "DATE_SUB(CONVERT_TZ(??, @@GLOBAL.time_zone, '+00:00'), INTERVAL SECOND(??) SECOND) as ??",
-                    [columnName, columnName, getAs(column)],
-                  ),
-                );
-              } else if (this.dbDriver.clientType() === 'sqlite3') {
-                selectors.push(
-                  this.dbDriver.raw(
-                    `strftime ('%Y-%m-%d %H:%M:00',:column:) ||
-  (
-   CASE WHEN substr(:column:, 20, 1) = '+' THEN
-    printf ('+%s:',
-     substr(:column:, 21, 2)) || printf ('%s',
-     substr(:column:, 24, 2))
-   WHEN substr(:column:, 20, 1) = '-' THEN
-    printf ('-%s:',
-     substr(:column:, 21, 2)) || printf ('%s',
-     substr(:column:, 24, 2))
-   ELSE
-    '+00:00'
-   END) AS :id:`,
-                    {
-                      column: columnName,
-                      id: getAs(column),
-                    },
-                  ),
-                );
-              } else {
-                selectors.push(
-                  this.dbDriver.raw('DATE(??) as ??', [
-                    columnName,
-                    getAs(column),
-                  ]),
-                );
-              }
-              groupBySelectors.push(getAs(column));
-            }
-            break;
-          default:
-            {
-              const columnName = await getColumnName(
-                this.context,
-                column,
-                columns,
-              );
-              selectors.push(
-                this.dbDriver.raw('?? as ??', [columnName, getAs(column)]),
-              );
-              groupBySelectors.push(getAs(column));
-            }
-            break;
-        }
-      }),
-    );
-
-    const qb = this.dbDriver(this.tnPath);
-
-    // get aggregated count of each group
-    qb.count(`${this.model.primaryKey?.column_name || '*'} as count`);
-
-    // get each group
-    qb.select(...selectors);
-
-    if (+rest?.shuffle) {
-      await this.shuffle({ qb });
-    }
-
-    const aliasColObjMap = await this.model.getAliasColObjMap(
-      this.context,
-      columns,
-    );
-
-    let sorts = extractSortsObject(rest?.sort, aliasColObjMap);
-
-    const { filters: filterObj } = extractFilterFromXwhere(
-      this.context,
-      where,
-      aliasColObjMap,
-    );
-    await conditionV2(
-      this,
-      [
-        ...(this.viewId
-          ? [
-              new Filter({
-                children:
-                  (await Filter.rootFilterList(this.context, {
-                    viewId: this.viewId,
-                  })) || [],
-                is_group: true,
-              }),
-            ]
-          : []),
-        new Filter({
-          children: args.filterArr || [],
-          is_group: true,
-          logical_op: 'and',
-        }),
-        new Filter({
-          children: filterObj,
-          is_group: true,
-          logical_op: 'and',
-        }),
-      ],
-      qb,
-    );
-
-    if (!sorts) {
-      if (args.sortArr?.length) {
-        sorts = args.sortArr;
-      } else if (this.viewId) {
-        sorts = await Sort.list(this.context, { viewId: this.viewId });
-      }
-    }
-
-    // if sort is provided filter out the group by columns sort and apply
-    // since we are grouping by the column and applying sort on any other column is not required
-    for (const sort of sorts || []) {
-      if (!groupByColumns[sort.fk_column_id]) {
-        continue;
-      }
-
-      const column = groupByColumns[sort.fk_column_id];
-
-      if (
-        [UITypes.User, UITypes.CreatedBy, UITypes.LastModifiedBy].includes(
-          column.uidt as UITypes,
-        )
-      ) {
-        const columnName = await getColumnName(this.context, column, columns);
-
-        const baseUsers = await BaseUser.getUsersList(this.context, {
-          base_id: column.base_id,
-        });
-
-        // create nested replace statement for each user
-        const finalStatement = baseUsers.reduce((acc, user) => {
-          const qb = this.dbDriver.raw(`REPLACE(${acc}, ?, ?)`, [
-            user.id,
-            user.display_name || user.email,
-          ]);
-          return qb.toQuery();
-        }, this.dbDriver.raw(`??`, [columnName]).toQuery());
-
-        if (!['asc', 'desc'].includes(sort.direction)) {
-          qb.orderBy(
-            'count',
-            sort.direction === 'count-desc' ? 'desc' : 'asc',
-            sort.direction === 'count-desc' ? 'LAST' : 'FIRST',
-          );
-        } else {
-          qb.orderBy(
-            sanitize(this.dbDriver.raw(finalStatement)),
-            sort.direction,
-            sort.direction === 'desc' ? 'LAST' : 'FIRST',
-          );
-        }
-      } else {
-        if (!['asc', 'desc'].includes(sort.direction)) {
-          qb.orderBy(
-            'count',
-            sort.direction === 'count-desc' ? 'desc' : 'asc',
-            sort.direction === 'count-desc' ? 'LAST' : 'FIRST',
-          );
-        } else {
-          qb.orderBy(
-            getAs(column),
-            sort.direction,
-            sort.direction === 'desc' ? 'LAST' : 'FIRST',
-          );
-        }
-      }
-    }
-
-    // group by using the column aliases
-    qb.groupBy(...groupBySelectors);
-
-    applyPaginate(qb, rest);
-
-    return await this.execAndParse(qb);
+    return await baseModelGroupBy(this, logger).list(args);
   }
 
   async groupByCount(args: {
@@ -2097,231 +1134,7 @@ class BaseModelSqlv2 implements IBaseModelSqlV2 {
     offset?;
     filterArr?: Filter[];
   }) {
-    const { where } = this._getListArgs(args as any);
-
-    args.column_name = args.column_name || '';
-
-    const selectors = [];
-    const groupBySelectors = [];
-    const getAlias = getAliasGenerator('__nc_gb');
-
-    const columns = await this.model.getColumns(this.context);
-
-    // todo: refactor and avoid duplicate code
-    await Promise.all(
-      args.column_name.split(',').map(async (col) => {
-        let column = columns.find(
-          (c) => c.column_name === col || c.title === col,
-        );
-        if (!column) {
-          throw NcError.fieldNotFound(col);
-        }
-
-        // if qrCode or Barcode replace it with value column nd keep the alias
-        if ([UITypes.QrCode, UITypes.Barcode].includes(column.uidt))
-          column = new Column({
-            ...(await column
-              .getColOptions<BarcodeColumn | QrCodeColumn>(this.context)
-              .then((col) => col.getValueColumn(this.context))),
-            asId: column.id,
-          });
-
-        switch (column.uidt) {
-          case UITypes.Attachment:
-            NcError.badRequest(
-              'Group by using attachment column is not supported',
-            );
-            break;
-          case UITypes.Button: {
-            NcError.badRequest('Group by using Button column is not supported');
-            break;
-          }
-          case UITypes.Rollup:
-          case UITypes.Links:
-            selectors.push(
-              (
-                await genRollupSelectv2({
-                  baseModelSqlv2: this,
-                  // tn: this.title,
-                  knex: this.dbDriver,
-                  // column,
-                  // alias,
-                  columnOptions: (await column.getColOptions(
-                    this.context,
-                  )) as RollupColumn,
-                })
-              ).builder.as(getAs(column)),
-            );
-            groupBySelectors.push(getAs(column));
-            break;
-          case UITypes.Formula: {
-            let selectQb;
-            try {
-              const _selectQb = await this.getSelectQueryBuilderForFormula(
-                column,
-              );
-
-              selectQb = this.dbDriver.raw(`?? as ??`, [
-                _selectQb.builder,
-                getAs(column),
-              ]);
-            } catch (e) {
-              logger.log(e);
-              // return dummy select
-              selectQb = this.dbDriver.raw(`'ERR' as ??`, [getAs(column)]);
-            }
-
-            selectors.push(selectQb);
-            groupBySelectors.push(getAs(column));
-            break;
-          }
-          case UITypes.Lookup:
-          case UITypes.LinkToAnotherRecord:
-            {
-              const _selectQb = await generateLookupSelectQuery({
-                baseModelSqlv2: this,
-                column,
-                alias: null,
-                model: this.model,
-                getAlias,
-              });
-
-              const selectQb = this.dbDriver.raw(`?? as ??`, [
-                this.dbDriver.raw(_selectQb.builder).wrap('(', ')'),
-                getAs(column),
-              ]);
-
-              selectors.push(selectQb);
-              groupBySelectors.push(getAs(column));
-            }
-            break;
-          case UITypes.CreatedTime:
-          case UITypes.LastModifiedTime:
-          case UITypes.DateTime:
-            {
-              const columnName = await getColumnName(
-                this.context,
-                column,
-                columns,
-              );
-              // ignore seconds part in datetime and group
-              if (this.dbDriver.clientType() === 'pg') {
-                selectors.push(
-                  this.dbDriver.raw(
-                    "date_trunc('minute', ??) + interval '0 seconds' as ??",
-                    [columnName, getAs(column)],
-                  ),
-                );
-              } else if (
-                this.dbDriver.clientType() === 'mysql' ||
-                this.dbDriver.clientType() === 'mysql2'
-              ) {
-                selectors.push(
-                  this.dbDriver.raw(
-                    "CONVERT_TZ(DATE_SUB(??, INTERVAL SECOND(??) SECOND), @@GLOBAL.time_zone, '+00:00') as ??",
-                    [columnName, columnName, getAs(column)],
-                  ),
-                );
-              } else if (this.dbDriver.clientType() === 'sqlite3') {
-                selectors.push(
-                  this.dbDriver.raw(
-                    `strftime ('%Y-%m-%d %H:%M:00',:column:) ||
-  (
-   CASE WHEN substr(:column:, 20, 1) = '+' THEN
-    printf ('+%s:',
-     substr(:column:, 21, 2)) || printf ('%s',
-     substr(:column:, 24, 2))
-   WHEN substr(:column:, 20, 1) = '-' THEN
-    printf ('-%s:',
-     substr(:column:, 21, 2)) || printf ('%s',
-     substr(:column:, 24, 2))
-   ELSE
-    '+00:00'
-   END) as :id:`,
-                    {
-                      column: columnName,
-                      id: getAs(column),
-                    },
-                  ),
-                );
-              } else {
-                selectors.push(
-                  this.dbDriver.raw('DATE(??) as ??', [
-                    columnName,
-                    getAs(column),
-                  ]),
-                );
-              }
-              groupBySelectors.push(getAs(column));
-            }
-            break;
-          default:
-            {
-              const columnName = await getColumnName(
-                this.context,
-                column,
-                columns,
-              );
-              selectors.push(
-                this.dbDriver.raw('?? as ??', [columnName, getAs(column)]),
-              );
-              groupBySelectors.push(getAs(column));
-            }
-            break;
-        }
-      }),
-    );
-
-    const qb = this.dbDriver(this.tnPath);
-    qb.count(`${this.model.primaryKey?.column_name || '*'} as count`);
-    qb.select(...selectors);
-
-    const aliasColObjMap = await this.model.getAliasColObjMap(
-      this.context,
-      columns,
-    );
-
-    const { filters: filterObj } = extractFilterFromXwhere(
-      this.context,
-      where,
-      aliasColObjMap,
-    );
-    await conditionV2(
-      this,
-      [
-        ...(this.viewId
-          ? [
-              new Filter({
-                children:
-                  (await Filter.rootFilterList(this.context, {
-                    viewId: this.viewId,
-                  })) || [],
-                is_group: true,
-              }),
-            ]
-          : []),
-        new Filter({
-          children: args.filterArr || [],
-          is_group: true,
-          logical_op: 'and',
-        }),
-        new Filter({
-          children: filterObj,
-          is_group: true,
-          logical_op: 'and',
-        }),
-      ],
-      qb,
-    );
-
-    qb.groupBy(...groupBySelectors);
-
-    const qbP = this.dbDriver
-      .count('*', { as: 'count' })
-      .from(qb.as('groupby'));
-
-    return (await this.execAndParse(qbP, null, { raw: true, first: true }))
-      ?.count;
+    return await baseModelGroupBy(this, logger).count(args);
   }
 
   // #region relation list count part 1
@@ -7703,800 +6516,22 @@ class BaseModelSqlv2 implements IBaseModelSqlV2 {
     return data;
   }
 
-  async addLinks({
-    cookie,
-    childIds: _childIds,
-    colId,
-    rowId,
-  }: {
+  async addLinks(params: {
     cookie: any;
     childIds: (string | number | Record<string, any>)[];
     colId: string;
     rowId: string;
   }) {
-    await this.model.getColumns(this.context);
-    const column = this.model.columnsById[colId];
-
-    if (!column || !isLinksOrLTAR(column)) NcError.fieldNotFound(colId);
-
-    const row = await this.readByPk(
-      rowId,
-      false,
-      {},
-      { ignoreView: true, getHiddenColumn: true },
-    );
-
-    // validate rowId
-    if (!row) {
-      NcError.recordNotFound(rowId);
-    }
-
-    if (!_childIds.length) return;
-
-    const colOptions = await column.getColOptions<LinkToAnotherRecordColumn>(
-      this.context,
-    );
-
-    const childColumn = await colOptions.getChildColumn(this.context);
-    const parentColumn = await colOptions.getParentColumn(this.context);
-    const parentTable = await parentColumn.getModel(this.context);
-    const childTable = await childColumn.getModel(this.context);
-    await childTable.getColumns(this.context);
-    await parentTable.getColumns(this.context);
-
-    const childBaseModel = await Model.getBaseModelSQL(this.context, {
-      model: childTable,
-      dbDriver: this.dbDriver,
-    });
-
-    const parentBaseModel = await Model.getBaseModelSQL(this.context, {
-      model: parentTable,
-      dbDriver: this.dbDriver,
-    });
-
-    const childTn = childBaseModel.getTnPath(childTable);
-    const parentTn = parentBaseModel.getTnPath(parentTable);
-
-    let relationType = colOptions.type;
-    let childIds = _childIds;
-
-    const relatedChildCol = getRelatedLinksColumn(
-      column,
-      this.model.id === parentTable.id ? childTable : parentTable,
-    );
-
-    const auditUpdateObj = [] as {
-      pkValue?: Record<string, any>;
-      columnTitle: string;
-      columnId: string;
-      refColumnTitle?: string;
-      rowId: unknown;
-      refRowId?: unknown;
-      req: NcRequest;
-      model: Model;
-      refModel?: Model;
-      displayValue?: unknown;
-      refDisplayValue?: unknown;
-      opSubType:
-        | AuditOperationSubTypes.LINK_RECORD
-        | AuditOperationSubTypes.UNLINK_RECORD;
-      type: RelationTypes;
-    }[];
-
-    const auditConfig = {
-      childModel: childTable,
-      parentModel: parentTable,
-      childColTitle: relatedChildCol?.title || '',
-      parentColTitle: column.title,
-      childColId: relatedChildCol?.id || '',
-      parentColId: column.id,
-    } as {
-      childModel: Model;
-      parentModel: Model;
-      childColTitle: string;
-      parentColTitle: string;
-      childColId: string;
-      parentColId: string;
-    };
-
-    if (relationType === RelationTypes.ONE_TO_ONE) {
-      relationType = column.meta?.bt
-        ? RelationTypes.BELONGS_TO
-        : RelationTypes.HAS_MANY;
-      childIds = childIds.slice(0, 1);
-
-      // unlink
-      await this.execAndParse(
-        this.dbDriver(childTn)
-          .where({
-            [childColumn.column_name]: this.dbDriver.from(
-              this.dbDriver(parentTn)
-                .select(parentColumn.column_name)
-                .where(
-                  _wherePk(
-                    parentTable.primaryKeys,
-                    column.meta?.bt ? childIds[0] : rowId,
-                  ),
-                )
-                .first()
-                .as('___cn_alias'),
-            ),
-          })
-          .update({ [childColumn.column_name]: null }),
-        null,
-        { raw: true },
-      );
-
-      auditConfig.parentModel = column.meta?.bt ? childTable : parentTable;
-      auditConfig.childModel = column.meta?.bt ? parentTable : childTable;
-    }
-
-    switch (relationType) {
-      case RelationTypes.MANY_TO_MANY:
-        {
-          const vChildCol = await colOptions.getMMChildColumn(this.context);
-          const vParentCol = await colOptions.getMMParentColumn(this.context);
-          const vTable = await colOptions.getMMModel(this.context);
-
-          const assocBaseModel = await Model.getBaseModelSQL(this.context, {
-            model: vTable,
-            dbDriver: this.dbDriver,
-          });
-
-          const vTn = assocBaseModel.getTnPath(vTable);
-
-          let insertData: Record<string, any>[];
-
-          // validate Ids
-          {
-            const childRowsQb = this.dbDriver(parentTn)
-              .select(`${parentTable.table_name}.${parentColumn.column_name}`)
-              .select(`${vTable.table_name}.${vChildCol.column_name}`)
-              .leftJoin(vTn, (qb) => {
-                qb.on(
-                  `${vTable.table_name}.${vParentCol.column_name}`,
-                  `${parentTable.table_name}.${parentColumn.column_name}`,
-                ).andOn(
-                  `${vTable.table_name}.${vChildCol.column_name}`,
-                  this.dbDriver.raw('?', [
-                    row[childColumn.title] ?? row[childColumn.column_name],
-                  ]),
-                );
-              });
-
-            if (parentTable.primaryKeys.length > 1) {
-              childRowsQb.where((qb) => {
-                for (const childId of childIds) {
-                  qb.orWhere(_wherePk(parentTable.primaryKeys, childId));
-                }
-              });
-            } else {
-              childRowsQb.whereIn(
-                `${parentTable.table_name}.${parentTable.primaryKey.column_name}`,
-                typeof childIds[0] === 'object'
-                  ? childIds.map(
-                      (c) =>
-                        c[parentTable.primaryKey.title] ??
-                        c[parentTable.primaryKey.column_name],
-                    )
-                  : childIds,
-              );
-            }
-
-            if (parentTable.primaryKey.column_name !== parentColumn.column_name)
-              childRowsQb.select(
-                `${parentTable.table_name}.${parentTable.primaryKey.column_name}`,
-              );
-
-            const childRows = await this.execAndParse(childRowsQb, null, {
-              raw: true,
-            });
-
-            if (childRows.length !== childIds.length) {
-              const missingIds = childIds.filter(
-                (id) =>
-                  !childRows.find((r) => r[parentColumn.column_name] === id),
-              );
-
-              NcError.recordNotFound(extractIds(missingIds));
-            }
-
-            insertData = childRows
-              // skip existing links
-              .filter((childRow) => !childRow[vChildCol.column_name])
-              // generate insert data for new links
-              .map((childRow) => ({
-                [vParentCol.column_name]:
-                  childRow[parentColumn.title] ??
-                  childRow[parentColumn.column_name],
-                [vChildCol.column_name]:
-                  row[childColumn.title] ?? row[childColumn.column_name],
-              }));
-
-            // if no new links, return true
-            if (!insertData.length) return true;
-          }
-
-          // todo: use bulk insert
-          await this.execAndParse(this.dbDriver(vTn).insert(insertData), null, {
-            raw: true,
-          });
-
-          await this.updateLastModified({
-            model: parentTable,
-            rowIds: childIds,
-            cookie,
-          });
-
-          await this.updateLastModified({
-            model: childTable,
-            rowIds: [rowId],
-            cookie,
-          });
-
-          auditConfig.parentModel =
-            this.model.id === parentTable.id ? parentTable : childTable;
-          auditConfig.childModel =
-            this.model.id === parentTable.id ? childTable : parentTable;
-        }
-        break;
-      case RelationTypes.HAS_MANY:
-        {
-          // validate Ids
-          {
-            const childRowsQb = this.dbDriver(childTn).select(
-              childTable.primaryKey.column_name,
-            );
-
-            if (childTable.primaryKeys.length > 1) {
-              childRowsQb.where((qb) => {
-                for (const childId of childIds) {
-                  qb.orWhere(_wherePk(childTable.primaryKeys, childId));
-                }
-              });
-            } else {
-              childRowsQb.whereIn(
-                parentTable.primaryKey.column_name,
-                typeof childIds[0] === 'object'
-                  ? childIds.map(
-                      (c) =>
-                        c[parentTable.primaryKey.title] ??
-                        c[parentTable.primaryKey.column_name],
-                    )
-                  : childIds,
-              );
-            }
-
-            const childRows = await this.execAndParse(childRowsQb, null, {
-              raw: true,
-            });
-
-            if (childRows.length !== childIds.length) {
-              const missingIds = childIds.filter(
-                (id) =>
-                  !childRows.find((r) => r[parentColumn.column_name] === id),
-              );
-
-              NcError.recordNotFound(extractIds(missingIds));
-            }
-          }
-          const updateQb = this.dbDriver(childTn).update({
-            [childColumn.column_name]: this.dbDriver.from(
-              this.dbDriver(parentTn)
-                .select(parentColumn.column_name)
-                .where(_wherePk(parentTable.primaryKeys, rowId))
-                .first()
-                .as('___cn_alias'),
-            ),
-          });
-          if (childTable.primaryKeys.length > 1) {
-            updateQb.where((qb) => {
-              for (const childId of childIds) {
-                qb.orWhere(_wherePk(childTable.primaryKeys, childId));
-              }
-            });
-          } else {
-            updateQb.whereIn(
-              childTable.primaryKey.column_name,
-              typeof childIds[0] === 'object'
-                ? childIds.map(
-                    (c) =>
-                      c[childTable.primaryKey.title] ??
-                      c[childTable.primaryKey.column_name],
-                  )
-                : childIds,
-            );
-          }
-          await this.execAndParse(updateQb, null, { raw: true });
-
-          await this.updateLastModified({
-            model: parentTable,
-            rowIds: [rowId],
-            cookie,
-          });
-        }
-        break;
-      case RelationTypes.BELONGS_TO:
-        {
-          auditConfig.parentModel = childTable;
-          auditConfig.childModel = parentTable;
-
-          // validate Ids
-          {
-            const childRowsQb = this.dbDriver(parentTn)
-              .select(parentTable.primaryKey.column_name)
-              .where(_wherePk(parentTable.primaryKeys, childIds[0]))
-              .first();
-
-            const childRow = await this.execAndParse(childRowsQb, null, {
-              first: true,
-              raw: true,
-            });
-
-            if (!childRow) {
-              NcError.recordNotFound(extractIds(childIds, true));
-            }
-          }
-
-          await this.execAndParse(
-            this.dbDriver(childTn)
-              .update({
-                [childColumn.column_name]: this.dbDriver.from(
-                  this.dbDriver(parentTn)
-                    .select(parentColumn.column_name)
-                    .where(_wherePk(parentTable.primaryKeys, childIds[0]))
-                    // .whereIn(parentTable.primaryKey.column_name, childIds)
-                    .first()
-                    .as('___cn_alias'),
-                ),
-              })
-              .where(_wherePk(childTable.primaryKeys, rowId)),
-            null,
-            { raw: true },
-          );
-
-          await this.updateLastModified({
-            model: parentTable,
-            rowIds: [rowId],
-            cookie,
-          });
-        }
-        break;
-    }
-
-    for (const childId of childIds) {
-      const _childId =
-        typeof childId === 'object'
-          ? Object.values(childId).join('_')
-          : childId;
-
-      auditUpdateObj.push({
-        model: auditConfig.parentModel,
-        refModel: auditConfig.childModel,
-        rowId,
-        refRowId: _childId,
-        opSubType: AuditOperationSubTypes.LINK_RECORD,
-        columnTitle: auditConfig.parentColTitle,
-        columnId: auditConfig.parentColId,
-        req: cookie,
-        type: colOptions.type as RelationTypes,
-      });
-
-      if (parentTable.id !== childTable.id) {
-        auditUpdateObj.push({
-          model: auditConfig.childModel,
-          refModel: auditConfig.parentModel,
-          rowId: _childId,
-          refRowId: rowId,
-          opSubType: AuditOperationSubTypes.LINK_RECORD,
-          columnTitle: auditConfig.childColTitle,
-          columnId: auditConfig.childColId,
-          req: cookie,
-          type: getOppositeRelationType(colOptions.type),
-        });
-      }
-    }
-
-    await Promise.allSettled(
-      auditUpdateObj.map(async (updateObj) => {
-        await this.afterAddChild({
-          columnTitle: updateObj.columnTitle,
-          columnId: updateObj.columnId,
-          refColumnTitle: updateObj.refColumnTitle,
-          rowId: updateObj.rowId,
-          refRowId: updateObj.refRowId,
-          req: updateObj.req,
-          model: updateObj.model,
-          refModel: updateObj.refModel,
-          displayValue: updateObj.displayValue,
-          refDisplayValue: updateObj.refDisplayValue,
-          type: updateObj.type,
-        });
-      }),
-    );
+    return addOrRemoveLinks(this).addLinks(params);
   }
 
-  async removeLinks({
-    cookie,
-    childIds,
-    colId,
-    rowId,
-  }: {
+  async removeLinks(params: {
     cookie: any;
     childIds: (string | number | Record<string, any>)[];
     colId: string;
     rowId: string;
   }) {
-    await this.model.getColumns(this.context);
-    const column = this.model.columnsById[colId];
-
-    if (!column || !isLinksOrLTAR(column)) NcError.fieldNotFound(colId);
-
-    const row = await this.readByPk(
-      rowId,
-      false,
-      {},
-      { ignoreView: true, getHiddenColumn: true },
-    );
-
-    // validate rowId
-    if (!row) {
-      NcError.recordNotFound(rowId);
-    }
-
-    if (!childIds.length) return;
-
-    const colOptions = await column.getColOptions<LinkToAnotherRecordColumn>(
-      this.context,
-    );
-
-    const childColumn = await colOptions.getChildColumn(this.context);
-    const parentColumn = await colOptions.getParentColumn(this.context);
-    const parentTable = await parentColumn.getModel(this.context);
-    const childTable = await childColumn.getModel(this.context);
-    await childTable.getColumns(this.context);
-    await parentTable.getColumns(this.context);
-
-    const childBaseModel = await Model.getBaseModelSQL(this.context, {
-      model: childTable,
-      dbDriver: this.dbDriver,
-    });
-
-    const parentBaseModel = await Model.getBaseModelSQL(this.context, {
-      model: parentTable,
-      dbDriver: this.dbDriver,
-    });
-
-    const childTn = childBaseModel.getTnPath(childTable);
-    const parentTn = parentBaseModel.getTnPath(parentTable);
-
-    const relatedChildCol = getRelatedLinksColumn(
-      column,
-      this.model.id === parentTable.id ? childTable : parentTable,
-    );
-
-    const auditUpdateObj = [] as {
-      columnTitle: string;
-      columnId: string;
-      refColumnTitle?: string;
-      rowId: unknown;
-      refRowId: unknown;
-      req: NcRequest;
-      model: Model;
-      refModel?: Model;
-      displayValue?: unknown;
-      refDisplayValue?: unknown;
-      opSubType:
-        | AuditOperationSubTypes.LINK_RECORD
-        | AuditOperationSubTypes.UNLINK_RECORD;
-      type: RelationTypes;
-    }[];
-
-    const auditConfig = {
-      childModel: childTable,
-      parentModel: parentTable,
-      childColTitle: relatedChildCol?.title || '',
-      parentColTitle: column.title,
-      childColId: relatedChildCol?.id || '',
-      parentColId: column.id,
-    } as {
-      childModel: Model;
-      parentModel: Model;
-      childColTitle: string;
-      childColId: string;
-      parentColTitle: string;
-      parentColId: string;
-    };
-
-    switch (colOptions.type) {
-      case RelationTypes.MANY_TO_MANY:
-        {
-          const vChildCol = await colOptions.getMMChildColumn(this.context);
-          const vParentCol = await colOptions.getMMParentColumn(this.context);
-          const vTable = await colOptions.getMMModel(this.context);
-
-          const assocBaseModel = await Model.getBaseModelSQL(this.context, {
-            model: vTable,
-            dbDriver: this.dbDriver,
-          });
-
-          // validate Ids
-          {
-            const childRowsQb = this.dbDriver(parentTn).select(
-              parentColumn.column_name,
-            );
-
-            if (parentTable.primaryKeys.length > 1) {
-              childRowsQb.where((qb) => {
-                for (const childId of childIds) {
-                  qb.orWhere(_wherePk(parentTable.primaryKeys, childId));
-                }
-              });
-            } else if (typeof childIds[0] === 'object') {
-              childRowsQb.whereIn(
-                `${parentTable.table_name}.${parentTable.primaryKey.column_name}`,
-                childIds.map(
-                  (c) =>
-                    c[parentTable.primaryKey.title] ||
-                    c[parentTable.primaryKey.column_name],
-                ),
-              );
-            } else {
-              childRowsQb.whereIn(
-                `${parentTable.table_name}.${parentTable.primaryKey.column_name}`,
-                childIds,
-              );
-            }
-
-            if (parentTable.primaryKey.column_name !== parentColumn.column_name)
-              childRowsQb.select(
-                `${parentTable.table_name}.${parentTable.primaryKey.column_name}`,
-              );
-
-            const childRows = await this.execAndParse(childRowsQb, null, {
-              raw: true,
-            });
-
-            if (childRows.length !== childIds.length) {
-              const missingIds = childIds.filter(
-                (id) =>
-                  !childRows.find(
-                    (r) =>
-                      r[parentColumn.column_name] ===
-                      (typeof id === 'object'
-                        ? id[parentTable.primaryKey.title] ??
-                          id[parentTable.primaryKey.column_name]
-                        : id),
-                  ),
-              );
-
-              NcError.recordNotFound(extractIds(missingIds));
-            }
-          }
-
-          const vTn = assocBaseModel.getTnPath(vTable);
-
-          const delQb = this.dbDriver(vTn)
-            .where({
-              [vChildCol.column_name]: this.dbDriver(childTn)
-                .select(childColumn.column_name)
-                .where(_wherePk(childTable.primaryKeys, rowId))
-                .first(),
-            })
-            .delete();
-
-          delQb.whereIn(
-            `${vTable.table_name}.${vParentCol.column_name}`,
-            typeof childIds[0] === 'object'
-              ? childIds.map(
-                  (c) =>
-                    c[parentTable.primaryKey.title] ??
-                    c[parentTable.primaryKey.column_name],
-                )
-              : childIds,
-          );
-          await this.execAndParse(delQb, null, { raw: true });
-
-          await this.updateLastModified({
-            model: parentTable,
-            rowIds: childIds,
-            cookie,
-          });
-          await this.updateLastModified({
-            model: childTable,
-            rowIds: [rowId],
-            cookie,
-          });
-
-          auditConfig.parentModel =
-            this.model.id === parentTable.id ? parentTable : childTable;
-          auditConfig.childModel =
-            this.model.id === parentTable.id ? childTable : parentTable;
-        }
-        break;
-      case RelationTypes.HAS_MANY:
-        {
-          // validate Ids
-          {
-            const childRowsQb = this.dbDriver(childTn).select(
-              childTable.primaryKey.column_name,
-            );
-
-            if (parentTable.primaryKeys.length > 1) {
-              childRowsQb.where((qb) => {
-                for (const childId of childIds) {
-                  qb.orWhere(_wherePk(parentTable.primaryKeys, childId));
-                }
-              });
-            } else if (typeof childIds[0] === 'object') {
-              childRowsQb.whereIn(
-                parentTable.primaryKey.column_name,
-                childIds.map(
-                  (c) =>
-                    c[parentTable.primaryKey.title] ??
-                    c[parentTable.primaryKey.column_name],
-                ),
-              );
-            } else {
-              childRowsQb.whereIn(parentTable.primaryKey.column_name, childIds);
-            }
-
-            const childRows = await this.execAndParse(childRowsQb, null, {
-              raw: true,
-            });
-
-            if (childRows.length !== childIds.length) {
-              const missingIds = childIds.filter(
-                (id) =>
-                  !childRows.find(
-                    (r) =>
-                      r[parentColumn.column_name] ===
-                      (typeof id === 'object'
-                        ? id[parentTable.primaryKey.title] ??
-                          id[parentTable.primaryKey.column_name]
-                        : id),
-                  ),
-              );
-
-              NcError.recordNotFound(extractIds(missingIds));
-            }
-          }
-
-          const childRowsQb = this.dbDriver(childTn);
-
-          if (parentTable.primaryKeys.length > 1) {
-            childRowsQb.where((qb) => {
-              for (const childId of childIds) {
-                qb.orWhere(_wherePk(parentTable.primaryKeys, childId));
-              }
-            });
-          } else {
-            childRowsQb.whereIn(
-              parentTable.primaryKey.column_name,
-              typeof childIds[0] === 'object'
-                ? childIds.map(
-                    (c) =>
-                      c[parentTable.primaryKey.title] ??
-                      c[parentTable.primaryKey.column_name],
-                  )
-                : childIds,
-            );
-          }
-
-          await this.execAndParse(
-            childRowsQb.update({ [childColumn.column_name]: null }),
-            null,
-            { raw: true },
-          );
-
-          await this.updateLastModified({
-            model: parentTable,
-            rowIds: [rowId],
-            cookie,
-          });
-        }
-        break;
-      case RelationTypes.BELONGS_TO:
-        {
-          auditConfig.parentModel = childTable;
-          auditConfig.childModel = parentTable;
-
-          // validate Ids
-          {
-            if (childIds.length > 1)
-              NcError.unprocessableEntity(
-                'Request must contain only one parent id',
-              );
-
-            const childRowsQb = this.dbDriver(parentTn)
-              .select(parentTable.primaryKey.column_name)
-              .where(_wherePk(parentTable.primaryKeys, childIds[0]))
-              .first();
-
-            const childRow = await this.execAndParse(childRowsQb, null, {
-              first: true,
-              raw: true,
-            });
-
-            if (!childRow) {
-              NcError.recordNotFound(extractIds(childIds, true));
-            }
-          }
-
-          await this.execAndParse(
-            this.dbDriver(childTn)
-              // .where({
-              //   [childColumn.cn]: this.dbDriver(parentTable.tn)
-              //     .select(parentColumn.cn)
-              //     .where(parentTable.primaryKey.cn, childId)
-              //     .first()
-              // })
-              // .where(_wherePk(childTable.primaryKeys, rowId))
-              .where(childTable.primaryKey.column_name, rowId)
-              .update({ [childColumn.column_name]: null }),
-            null,
-            { raw: true },
-          );
-
-          await this.updateLastModified({
-            model: parentTable,
-            rowIds: [childIds[0]],
-            cookie,
-          });
-        }
-        break;
-    }
-
-    for (const childId of childIds) {
-      const _childId =
-        typeof childId === 'object'
-          ? Object.values(childId).join('_')
-          : childId;
-
-      auditUpdateObj.push({
-        model: auditConfig.parentModel,
-        refModel: auditConfig.childModel,
-        rowId,
-        refRowId: _childId,
-        opSubType: AuditOperationSubTypes.LINK_RECORD,
-        columnTitle: auditConfig.parentColTitle,
-        columnId: auditConfig.parentColId,
-        req: cookie,
-        type: colOptions.type as RelationTypes,
-      });
-
-      if (parentTable.id !== childTable.id) {
-        auditUpdateObj.push({
-          model: auditConfig.childModel,
-          refModel: auditConfig.parentModel,
-          rowId: _childId,
-          refRowId: rowId,
-          opSubType: AuditOperationSubTypes.LINK_RECORD,
-          columnTitle: auditConfig.childColTitle,
-          columnId: auditConfig.childColId,
-          req: cookie,
-          type: getOppositeRelationType(colOptions.type),
-        });
-      }
-    }
-
-    await Promise.allSettled(
-      auditUpdateObj.map(async (updateObj) => {
-        await this.afterRemoveChild({
-          columnTitle: updateObj.columnTitle,
-          columnId: updateObj.columnId,
-          refColumnTitle: updateObj.refColumnTitle,
-          rowId: updateObj.rowId,
-          refRowId: updateObj.refRowId,
-          req: updateObj.req,
-          model: updateObj.model,
-          refModel: updateObj.refModel,
-          displayValue: updateObj.displayValue,
-          refDisplayValue: updateObj.refDisplayValue,
-          type: updateObj.type,
-        });
-      }),
-    );
+    return addOrRemoveLinks(this).removeLinks(params);
   }
 
   async btRead(
