+++ packages/nocodb/src/db/BaseModelSqlv2/add-remove-links.ts
@@ -1,0 +1,847 @@
+import {
+  AuditOperationSubTypes,
+  isLinksOrLTAR,
+  RelationTypes,
+} from 'nocodb-sdk';
+import type { NcRequest } from 'nocodb-sdk';
+import type { IBaseModelSqlV2 } from '~/db/IBaseModelSqlV2';
+import type { LinkToAnotherRecordColumn } from '~/models';
+import { NcError } from '~/helpers/catchError';
+import {
+  _wherePk,
+  extractIds,
+  getOppositeRelationType,
+  getRelatedLinksColumn,
+} from '~/helpers/dbHelpers';
+import { Model } from '~/models';
+
+export const addOrRemoveLinks = (baseModel: IBaseModelSqlV2) => {
+  const addLinks = async ({
+    cookie,
+    childIds: _childIds,
+    colId,
+    rowId,
+  }: {
+    cookie: any;
+    childIds: (string | number | Record<string, any>)[];
+    colId: string;
+    rowId: string;
+  }) => {
+    await baseModel.model.getColumns(baseModel.context);
+    const column = baseModel.model.columnsById[colId];
+
+    if (!column || !isLinksOrLTAR(column)) NcError.fieldNotFound(colId);
+
+    const row = await baseModel.readByPk(
+      rowId,
+      false,
+      {},
+      { ignoreView: true, getHiddenColumn: true },
+    );
+
+    // validate rowId
+    if (!row) {
+      NcError.recordNotFound(rowId);
+    }
+
+    if (!_childIds.length) return;
+
+    const colOptions = await column.getColOptions<LinkToAnotherRecordColumn>(
+      baseModel.context,
+    );
+
+    const childColumn = await colOptions.getChildColumn(baseModel.context);
+    const parentColumn = await colOptions.getParentColumn(baseModel.context);
+    const parentTable = await parentColumn.getModel(baseModel.context);
+    const childTable = await childColumn.getModel(baseModel.context);
+    await childTable.getColumns(baseModel.context);
+    await parentTable.getColumns(baseModel.context);
+
+    const childBaseModel = await Model.getBaseModelSQL(baseModel.context, {
+      model: childTable,
+      dbDriver: baseModel.dbDriver,
+    });
+
+    const parentBaseModel = await Model.getBaseModelSQL(baseModel.context, {
+      model: parentTable,
+      dbDriver: baseModel.dbDriver,
+    });
+
+    const childTn = childBaseModel.getTnPath(childTable);
+    const parentTn = parentBaseModel.getTnPath(parentTable);
+
+    let relationType = colOptions.type;
+    let childIds = _childIds;
+
+    const relatedChildCol = getRelatedLinksColumn(
+      column,
+      baseModel.model.id === parentTable.id ? childTable : parentTable,
+    );
+
+    const auditUpdateObj = [] as {
+      pkValue?: Record<string, any>;
+      columnTitle: string;
+      columnId: string;
+      refColumnTitle?: string;
+      rowId: unknown;
+      refRowId?: unknown;
+      req: NcRequest;
+      model: Model;
+      refModel?: Model;
+      displayValue?: unknown;
+      refDisplayValue?: unknown;
+      opSubType:
+        | AuditOperationSubTypes.LINK_RECORD
+        | AuditOperationSubTypes.UNLINK_RECORD;
+      type: RelationTypes;
+    }[];
+
+    const auditConfig = {
+      childModel: childTable,
+      parentModel: parentTable,
+      childColTitle: relatedChildCol?.title || '',
+      parentColTitle: column.title,
+      childColId: relatedChildCol?.id || '',
+      parentColId: column.id,
+    } as {
+      childModel: Model;
+      parentModel: Model;
+      childColTitle: string;
+      parentColTitle: string;
+      childColId: string;
+      parentColId: string;
+    };
+
+    if (relationType === RelationTypes.ONE_TO_ONE) {
+      relationType = column.meta?.bt
+        ? RelationTypes.BELONGS_TO
+        : RelationTypes.HAS_MANY;
+      childIds = childIds.slice(0, 1);
+
+      // unlink
+      await baseModel.execAndParse(
+        baseModel
+          .dbDriver(childTn)
+          .where({
+            [childColumn.column_name]: baseModel.dbDriver.from(
+              baseModel
+                .dbDriver(parentTn)
+                .select(parentColumn.column_name)
+                .where(
+                  _wherePk(
+                    parentTable.primaryKeys,
+                    column.meta?.bt ? childIds[0] : rowId,
+                  ),
+                )
+                .first()
+                .as('___cn_alias'),
+            ),
+          })
+          .update({ [childColumn.column_name]: null }),
+        null,
+        { raw: true },
+      );
+
+      auditConfig.parentModel = column.meta?.bt ? childTable : parentTable;
+      auditConfig.childModel = column.meta?.bt ? parentTable : childTable;
+    }
+
+    switch (relationType) {
+      case RelationTypes.MANY_TO_MANY:
+        {
+          const vChildCol = await colOptions.getMMChildColumn(
+            baseModel.context,
+          );
+          const vParentCol = await colOptions.getMMParentColumn(
+            baseModel.context,
+          );
+          const vTable = await colOptions.getMMModel(baseModel.context);
+
+          const assocBaseModel = await Model.getBaseModelSQL(
+            baseModel.context,
+            {
+              model: vTable,
+              dbDriver: baseModel.dbDriver,
+            },
+          );
+
+          const vTn = assocBaseModel.getTnPath(vTable);
+
+          let insertData: Record<string, any>[];
+
+          // validate Ids
+          {
+            const childRowsQb = baseModel
+              .dbDriver(parentTn)
+              .select(`${parentTable.table_name}.${parentColumn.column_name}`)
+              .select(`${vTable.table_name}.${vChildCol.column_name}`)
+              .leftJoin(vTn, (qb) => {
+                qb.on(
+                  `${vTable.table_name}.${vParentCol.column_name}`,
+                  `${parentTable.table_name}.${parentColumn.column_name}`,
+                ).andOn(
+                  `${vTable.table_name}.${vChildCol.column_name}`,
+                  baseModel.dbDriver.raw('?', [
+                    row[childColumn.title] ?? row[childColumn.column_name],
+                  ]),
+                );
+              });
+
+            if (parentTable.primaryKeys.length > 1) {
+              childRowsQb.where((qb) => {
+                for (const childId of childIds) {
+                  qb.orWhere(_wherePk(parentTable.primaryKeys, childId));
+                }
+              });
+            } else {
+              childRowsQb.whereIn(
+                `${parentTable.table_name}.${parentTable.primaryKey.column_name}`,
+                typeof childIds[0] === 'object'
+                  ? childIds.map(
+                      (c) =>
+                        c[parentTable.primaryKey.title] ??
+                        c[parentTable.primaryKey.column_name],
+                    )
+                  : childIds,
+              );
+            }
+
+            if (parentTable.primaryKey.column_name !== parentColumn.column_name)
+              childRowsQb.select(
+                `${parentTable.table_name}.${parentTable.primaryKey.column_name}`,
+              );
+
+            const childRows = await baseModel.execAndParse(childRowsQb, null, {
+              raw: true,
+            });
+
+            if (childRows.length !== childIds.length) {
+              const missingIds = childIds.filter(
+                (id) =>
+                  !childRows.find((r) => r[parentColumn.column_name] === id),
+              );
+
+              NcError.recordNotFound(extractIds(missingIds));
+            }
+
+            insertData = childRows
+              // skip existing links
+              .filter((childRow) => !childRow[vChildCol.column_name])
+              // generate insert data for new links
+              .map((childRow) => ({
+                [vParentCol.column_name]:
+                  childRow[parentColumn.title] ??
+                  childRow[parentColumn.column_name],
+                [vChildCol.column_name]:
+                  row[childColumn.title] ?? row[childColumn.column_name],
+              }));
+
+            // if no new links, return true
+            if (!insertData.length) return true;
+          }
+
+          // todo: use bulk insert
+          await baseModel.execAndParse(
+            baseModel.dbDriver(vTn).insert(insertData),
+            null,
+            {
+              raw: true,
+            },
+          );
+
+          await baseModel.updateLastModified({
+            model: parentTable,
+            rowIds: childIds,
+            cookie,
+          });
+
+          await baseModel.updateLastModified({
+            model: childTable,
+            rowIds: [rowId],
+            cookie,
+          });
+
+          auditConfig.parentModel =
+            baseModel.model.id === parentTable.id ? parentTable : childTable;
+          auditConfig.childModel =
+            baseModel.model.id === parentTable.id ? childTable : parentTable;
+        }
+        break;
+      case RelationTypes.HAS_MANY:
+        {
+          // validate Ids
+          {
+            const childRowsQb = baseModel
+              .dbDriver(childTn)
+              .select(childTable.primaryKey.column_name);
+
+            if (childTable.primaryKeys.length > 1) {
+              childRowsQb.where((qb) => {
+                for (const childId of childIds) {
+                  qb.orWhere(_wherePk(childTable.primaryKeys, childId));
+                }
+              });
+            } else {
+              childRowsQb.whereIn(
+                parentTable.primaryKey.column_name,
+                typeof childIds[0] === 'object'
+                  ? childIds.map(
+                      (c) =>
+                        c[parentTable.primaryKey.title] ??
+                        c[parentTable.primaryKey.column_name],
+                    )
+                  : childIds,
+              );
+            }
+
+            const childRows = await baseModel.execAndParse(childRowsQb, null, {
+              raw: true,
+            });
+
+            if (childRows.length !== childIds.length) {
+              const missingIds = childIds.filter(
+                (id) =>
+                  !childRows.find((r) => r[parentColumn.column_name] === id),
+              );
+
+              NcError.recordNotFound(extractIds(missingIds));
+            }
+          }
+          const updateQb = baseModel.dbDriver(childTn).update({
+            [childColumn.column_name]: baseModel.dbDriver.from(
+              baseModel
+                .dbDriver(parentTn)
+                .select(parentColumn.column_name)
+                .where(_wherePk(parentTable.primaryKeys, rowId))
+                .first()
+                .as('___cn_alias'),
+            ),
+          });
+          if (childTable.primaryKeys.length > 1) {
+            updateQb.where((qb) => {
+              for (const childId of childIds) {
+                qb.orWhere(_wherePk(childTable.primaryKeys, childId));
+              }
+            });
+          } else {
+            updateQb.whereIn(
+              childTable.primaryKey.column_name,
+              typeof childIds[0] === 'object'
+                ? childIds.map(
+                    (c) =>
+                      c[childTable.primaryKey.title] ??
+                      c[childTable.primaryKey.column_name],
+                  )
+                : childIds,
+            );
+          }
+          await baseModel.execAndParse(updateQb, null, { raw: true });
+
+          await baseModel.updateLastModified({
+            model: parentTable,
+            rowIds: [rowId],
+            cookie,
+          });
+        }
+        break;
+      case RelationTypes.BELONGS_TO:
+        {
+          auditConfig.parentModel = childTable;
+          auditConfig.childModel = parentTable;
+
+          // validate Ids
+          {
+            const childRowsQb = baseModel
+              .dbDriver(parentTn)
+              .select(parentTable.primaryKey.column_name)
+              .where(_wherePk(parentTable.primaryKeys, childIds[0]))
+              .first();
+
+            const childRow = await baseModel.execAndParse(childRowsQb, null, {
+              first: true,
+              raw: true,
+            });
+
+            if (!childRow) {
+              NcError.recordNotFound(extractIds(childIds, true));
+            }
+          }
+
+          await baseModel.execAndParse(
+            baseModel
+              .dbDriver(childTn)
+              .update({
+                [childColumn.column_name]: baseModel.dbDriver.from(
+                  baseModel
+                    .dbDriver(parentTn)
+                    .select(parentColumn.column_name)
+                    .where(_wherePk(parentTable.primaryKeys, childIds[0]))
+                    // .whereIn(parentTable.primaryKey.column_name, childIds)
+                    .first()
+                    .as('___cn_alias'),
+                ),
+              })
+              .where(_wherePk(childTable.primaryKeys, rowId)),
+            null,
+            { raw: true },
+          );
+
+          await baseModel.updateLastModified({
+            model: parentTable,
+            rowIds: [rowId],
+            cookie,
+          });
+        }
+        break;
+    }
+
+    for (const childId of childIds) {
+      const _childId =
+        typeof childId === 'object'
+          ? Object.values(childId).join('_')
+          : childId;
+
+      auditUpdateObj.push({
+        model: auditConfig.parentModel,
+        refModel: auditConfig.childModel,
+        rowId,
+        refRowId: _childId,
+        opSubType: AuditOperationSubTypes.LINK_RECORD,
+        columnTitle: auditConfig.parentColTitle,
+        columnId: auditConfig.parentColId,
+        req: cookie,
+        type: colOptions.type as RelationTypes,
+      });
+
+      if (parentTable.id !== childTable.id) {
+        auditUpdateObj.push({
+          model: auditConfig.childModel,
+          refModel: auditConfig.parentModel,
+          rowId: _childId,
+          refRowId: rowId,
+          opSubType: AuditOperationSubTypes.LINK_RECORD,
+          columnTitle: auditConfig.childColTitle,
+          columnId: auditConfig.childColId,
+          req: cookie,
+          type: getOppositeRelationType(colOptions.type),
+        });
+      }
+    }
+
+    await Promise.allSettled(
+      auditUpdateObj.map(async (updateObj) => {
+        await baseModel.afterAddChild({
+          columnTitle: updateObj.columnTitle,
+          columnId: updateObj.columnId,
+          refColumnTitle: updateObj.refColumnTitle,
+          rowId: updateObj.rowId,
+          refRowId: updateObj.refRowId,
+          req: updateObj.req,
+          model: updateObj.model,
+          refModel: updateObj.refModel,
+          displayValue: updateObj.displayValue,
+          refDisplayValue: updateObj.refDisplayValue,
+          type: updateObj.type,
+        });
+      }),
+    );
+  };
+
+  const removeLinks = async ({
+    cookie,
+    childIds,
+    colId,
+    rowId,
+  }: {
+    cookie: any;
+    childIds: (string | number | Record<string, any>)[];
+    colId: string;
+    rowId: string;
+  }) => {
+    await baseModel.model.getColumns(baseModel.context);
+    const column = baseModel.model.columnsById[colId];
+
+    if (!column || !isLinksOrLTAR(column)) NcError.fieldNotFound(colId);
+
+    const row = await baseModel.readByPk(
+      rowId,
+      false,
+      {},
+      { ignoreView: true, getHiddenColumn: true },
+    );
+
+    // validate rowId
+    if (!row) {
+      NcError.recordNotFound(rowId);
+    }
+
+    if (!childIds.length) return;
+
+    const colOptions = await column.getColOptions<LinkToAnotherRecordColumn>(
+      baseModel.context,
+    );
+
+    const childColumn = await colOptions.getChildColumn(baseModel.context);
+    const parentColumn = await colOptions.getParentColumn(baseModel.context);
+    const parentTable = await parentColumn.getModel(baseModel.context);
+    const childTable = await childColumn.getModel(baseModel.context);
+    await childTable.getColumns(baseModel.context);
+    await parentTable.getColumns(baseModel.context);
+
+    const childBaseModel = await Model.getBaseModelSQL(baseModel.context, {
+      model: childTable,
+      dbDriver: baseModel.dbDriver,
+    });
+
+    const parentBaseModel = await Model.getBaseModelSQL(baseModel.context, {
+      model: parentTable,
+      dbDriver: baseModel.dbDriver,
+    });
+
+    const childTn = childBaseModel.getTnPath(childTable);
+    const parentTn = parentBaseModel.getTnPath(parentTable);
+
+    const relatedChildCol = getRelatedLinksColumn(
+      column,
+      baseModel.model.id === parentTable.id ? childTable : parentTable,
+    );
+
+    const auditUpdateObj = [] as {
+      columnTitle: string;
+      columnId: string;
+      refColumnTitle?: string;
+      rowId: unknown;
+      refRowId: unknown;
+      req: NcRequest;
+      model: Model;
+      refModel?: Model;
+      displayValue?: unknown;
+      refDisplayValue?: unknown;
+      opSubType:
+        | AuditOperationSubTypes.LINK_RECORD
+        | AuditOperationSubTypes.UNLINK_RECORD;
+      type: RelationTypes;
+    }[];
+
+    const auditConfig = {
+      childModel: childTable,
+      parentModel: parentTable,
+      childColTitle: relatedChildCol?.title || '',
+      parentColTitle: column.title,
+      childColId: relatedChildCol?.id || '',
+      parentColId: column.id,
+    } as {
+      childModel: Model;
+      parentModel: Model;
+      childColTitle: string;
+      childColId: string;
+      parentColTitle: string;
+      parentColId: string;
+    };
+
+    switch (colOptions.type) {
+      case RelationTypes.MANY_TO_MANY:
+        {
+          const vChildCol = await colOptions.getMMChildColumn(
+            baseModel.context,
+          );
+          const vParentCol = await colOptions.getMMParentColumn(
+            baseModel.context,
+          );
+          const vTable = await colOptions.getMMModel(baseModel.context);
+
+          const assocBaseModel = await Model.getBaseModelSQL(
+            baseModel.context,
+            {
+              model: vTable,
+              dbDriver: baseModel.dbDriver,
+            },
+          );
+
+          // validate Ids
+          {
+            const childRowsQb = baseModel
+              .dbDriver(parentTn)
+              .select(parentColumn.column_name);
+
+            if (parentTable.primaryKeys.length > 1) {
+              childRowsQb.where((qb) => {
+                for (const childId of childIds) {
+                  qb.orWhere(_wherePk(parentTable.primaryKeys, childId));
+                }
+              });
+            } else if (typeof childIds[0] === 'object') {
+              childRowsQb.whereIn(
+                `${parentTable.table_name}.${parentTable.primaryKey.column_name}`,
+                childIds.map(
+                  (c) =>
+                    c[parentTable.primaryKey.title] ||
+                    c[parentTable.primaryKey.column_name],
+                ),
+              );
+            } else {
+              childRowsQb.whereIn(
+                `${parentTable.table_name}.${parentTable.primaryKey.column_name}`,
+                childIds,
+              );
+            }
+
+            if (parentTable.primaryKey.column_name !== parentColumn.column_name)
+              childRowsQb.select(
+                `${parentTable.table_name}.${parentTable.primaryKey.column_name}`,
+              );
+
+            const childRows = await baseModel.execAndParse(childRowsQb, null, {
+              raw: true,
+            });
+
+            if (childRows.length !== childIds.length) {
+              const missingIds = childIds.filter(
+                (id) =>
+                  !childRows.find(
+                    (r) =>
+                      r[parentColumn.column_name] ===
+                      (typeof id === 'object'
+                        ? id[parentTable.primaryKey.title] ??
+                          id[parentTable.primaryKey.column_name]
+                        : id),
+                  ),
+              );
+
+              NcError.recordNotFound(extractIds(missingIds));
+            }
+          }
+
+          const vTn = assocBaseModel.getTnPath(vTable);
+
+          const delQb = baseModel
+            .dbDriver(vTn)
+            .where({
+              [vChildCol.column_name]: baseModel
+                .dbDriver(childTn)
+                .select(childColumn.column_name)
+                .where(_wherePk(childTable.primaryKeys, rowId))
+                .first(),
+            })
+            .delete();
+
+          delQb.whereIn(
+            `${vTable.table_name}.${vParentCol.column_name}`,
+            typeof childIds[0] === 'object'
+              ? childIds.map(
+                  (c) =>
+                    c[parentTable.primaryKey.title] ??
+                    c[parentTable.primaryKey.column_name],
+                )
+              : childIds,
+          );
+          await baseModel.execAndParse(delQb, null, { raw: true });
+
+          await baseModel.updateLastModified({
+            model: parentTable,
+            rowIds: childIds,
+            cookie,
+          });
+          await baseModel.updateLastModified({
+            model: childTable,
+            rowIds: [rowId],
+            cookie,
+          });
+
+          auditConfig.parentModel =
+            baseModel.model.id === parentTable.id ? parentTable : childTable;
+          auditConfig.childModel =
+            baseModel.model.id === parentTable.id ? childTable : parentTable;
+        }
+        break;
+      case RelationTypes.HAS_MANY:
+        {
+          // validate Ids
+          {
+            const childRowsQb = baseModel
+              .dbDriver(childTn)
+              .select(childTable.primaryKey.column_name);
+
+            if (parentTable.primaryKeys.length > 1) {
+              childRowsQb.where((qb) => {
+                for (const childId of childIds) {
+                  qb.orWhere(_wherePk(parentTable.primaryKeys, childId));
+                }
+              });
+            } else if (typeof childIds[0] === 'object') {
+              childRowsQb.whereIn(
+                parentTable.primaryKey.column_name,
+                childIds.map(
+                  (c) =>
+                    c[parentTable.primaryKey.title] ??
+                    c[parentTable.primaryKey.column_name],
+                ),
+              );
+            } else {
+              childRowsQb.whereIn(parentTable.primaryKey.column_name, childIds);
+            }
+
+            const childRows = await baseModel.execAndParse(childRowsQb, null, {
+              raw: true,
+            });
+
+            if (childRows.length !== childIds.length) {
+              const missingIds = childIds.filter(
+                (id) =>
+                  !childRows.find(
+                    (r) =>
+                      r[parentColumn.column_name] ===
+                      (typeof id === 'object'
+                        ? id[parentTable.primaryKey.title] ??
+                          id[parentTable.primaryKey.column_name]
+                        : id),
+                  ),
+              );
+
+              NcError.recordNotFound(extractIds(missingIds));
+            }
+          }
+
+          const childRowsQb = baseModel.dbDriver(childTn);
+
+          if (parentTable.primaryKeys.length > 1) {
+            childRowsQb.where((qb) => {
+              for (const childId of childIds) {
+                qb.orWhere(_wherePk(parentTable.primaryKeys, childId));
+              }
+            });
+          } else {
+            childRowsQb.whereIn(
+              parentTable.primaryKey.column_name,
+              typeof childIds[0] === 'object'
+                ? childIds.map(
+                    (c) =>
+                      c[parentTable.primaryKey.title] ??
+                      c[parentTable.primaryKey.column_name],
+                  )
+                : childIds,
+            );
+          }
+
+          await baseModel.execAndParse(
+            childRowsQb.update({ [childColumn.column_name]: null }),
+            null,
+            { raw: true },
+          );
+
+          await baseModel.updateLastModified({
+            model: parentTable,
+            rowIds: [rowId],
+            cookie,
+          });
+        }
+        break;
+      case RelationTypes.BELONGS_TO:
+        {
+          auditConfig.parentModel = childTable;
+          auditConfig.childModel = parentTable;
+
+          // validate Ids
+          {
+            if (childIds.length > 1)
+              NcError.unprocessableEntity(
+                'Request must contain only one parent id',
+              );
+
+            const childRowsQb = baseModel
+              .dbDriver(parentTn)
+              .select(parentTable.primaryKey.column_name)
+              .where(_wherePk(parentTable.primaryKeys, childIds[0]))
+              .first();
+
+            const childRow = await baseModel.execAndParse(childRowsQb, null, {
+              first: true,
+              raw: true,
+            });
+
+            if (!childRow) {
+              NcError.recordNotFound(extractIds(childIds, true));
+            }
+          }
+
+          await baseModel.execAndParse(
+            baseModel
+              .dbDriver(childTn)
+              // .where({
+              //   [childColumn.cn]: baseModel.dbDriver(parentTable.tn)
+              //     .select(parentColumn.cn)
+              //     .where(parentTable.primaryKey.cn, childId)
+              //     .first()
+              // })
+              // .where(_wherePk(childTable.primaryKeys, rowId))
+              .where(childTable.primaryKey.column_name, rowId)
+              .update({ [childColumn.column_name]: null }),
+            null,
+            { raw: true },
+          );
+
+          await baseModel.updateLastModified({
+            model: parentTable,
+            rowIds: [childIds[0]],
+            cookie,
+          });
+        }
+        break;
+    }
+
+    for (const childId of childIds) {
+      const _childId =
+        typeof childId === 'object'
+          ? Object.values(childId).join('_')
+          : childId;
+
+      auditUpdateObj.push({
+        model: auditConfig.parentModel,
+        refModel: auditConfig.childModel,
+        rowId,
+        refRowId: _childId,
+        opSubType: AuditOperationSubTypes.LINK_RECORD,
+        columnTitle: auditConfig.parentColTitle,
+        columnId: auditConfig.parentColId,
+        req: cookie,
+        type: colOptions.type as RelationTypes,
+      });
+
+      if (parentTable.id !== childTable.id) {
+        auditUpdateObj.push({
+          model: auditConfig.childModel,
+          refModel: auditConfig.parentModel,
+          rowId: _childId,
+          refRowId: rowId,
+          opSubType: AuditOperationSubTypes.LINK_RECORD,
+          columnTitle: auditConfig.childColTitle,
+          columnId: auditConfig.childColId,
+          req: cookie,
+          type: getOppositeRelationType(colOptions.type),
+        });
+      }
+    }
+
+    await Promise.allSettled(
+      auditUpdateObj.map(async (updateObj) => {
+        await baseModel.afterRemoveChild({
+          columnTitle: updateObj.columnTitle,
+          columnId: updateObj.columnId,
+          refColumnTitle: updateObj.refColumnTitle,
+          rowId: updateObj.rowId,
+          refRowId: updateObj.refRowId,
+          req: updateObj.req,
+          model: updateObj.model,
+          refModel: updateObj.refModel,
+          displayValue: updateObj.displayValue,
+          refDisplayValue: updateObj.refDisplayValue,
+          type: updateObj.type,
+        });
+      }),
+    );
+  };
+  return {
+    addLinks,
+    removeLinks,
+  };
+};
